refactor : 우승자출력 시 사용되는 문장 만들기 함수들을 OutputView에서 RaceGame으로 이동
