+++ README.md
@@ -9,6 +9,9 @@
   - [x] 전진 관리 함수
   - [x] 난수 크기 확인함수
   - [x] 우승자 출력 함수
+	- [x] 가장 많은 이동 횟수를 반환하는 함수
+	- [x] 우승자 StringBuffer를 반환하는 함수
+	- [x] 우승자 출력 시 이름 사이에 쉼표를 출력하기 위한 함수 
 
 - RaceGameManager (게임 객체를 통해 게임을 실행하고 조종하는 클래스)
   - [x] 게임 시작 함수
@@ -35,9 +38,6 @@
   - [x] 에러메세지 출력
   - [x] 실행결과 출력
   - [x] 최종 우승자 출력
-	- [x] 가장 많은 이동 횟수를 반환하는 함수
-	- [x] 우승자 StringBuffer를 반환하는 함수
-	- [x] 우승자 출력 시 이름 사이에 쉼표를 출력하기 위한 함수 
 
 - NameValidator (이름 입력 유효성 검사를 담당하는 클래스)
