+++ src/main/java/racegame/view/OutputView.java
@@ -2,7 +2,6 @@ package racegame.view;
 
 import racingcar.Car;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import static racegame.Constant.*;
@@ -33,38 +32,8 @@ public class OutputView {
         System.out.println();
     }
 
-    public static void printWinners(List<Car> carList) {
+    public static void printWinners(String winners) {
         System.out.print(OUTPUT_VIEW_WINNER_HEADER);
-
-        int maxPosition = getMaximumPosition(carList);
-        StringBuffer winners = getWinnersStringBuffer(carList, maxPosition);
-
         System.out.println(winners);
-    }
-
-    private static int getMaximumPosition(List<Car> carList){
-        List<Integer> list = new ArrayList<>();
-        for (Car car : carList) {
-            list.add(car.getPosition());
-        }
-        return list.stream().max(Integer::compare).orElse(0);
-    }
-
-    private static StringBuffer getWinnersStringBuffer(List<Car> carList, int maxPosition){
-        StringBuffer winners = new StringBuffer();
-        for(Car car : carList){
-            if(car.getPosition() == maxPosition) {
-                winners = printNameDelimiter(winners, winners.length());
-                winners.append(car.getName());
-            }
-        }
-        return winners;
-    }
-
-    private static StringBuffer printNameDelimiter(StringBuffer winners, int length){
-        if(length >0){
-            winners.append(OUTPUT_VIEW_WINNER_DELIMITER);
-        }
-        return winners;
     }
 }
